图推理/requirements.txt
Flask==2.2.1
matplotlib==3.5.2
nebula3-python==3.1.0
networkx==2.8.5
numpy==1.23.1
scipy==1.9.0